Use dialog font for debug text

+++ Source/Headers/ogl_support.h
@@ -264,6 +264,6 @@ void OGL_PopState(void);
 void OGL_EnableLighting(void);
 void OGL_DisableLighting(void);
 
-void OGL_DrawString(const char* s, GLint x, GLint y);
-void OGL_DrawFloat(float f, GLint x, GLint y);
-void OGL_DrawInt(int f, GLint x, GLint y);
+void OGL_DrawString(const char* s, float x, float y);
+void OGL_DrawFloat(float f, float x, float y);
+void OGL_DrawInt(int f, float x, float y);
